Move internal err tolerance logic and types to supervisor package
