+++ CHANGELOG.md
@@ -1,0 +1,7 @@
+# pre-release v0.1.0 (breaking changes)
+
+* Deprecate `WithTolerance` worker option with `WithErrTolerance` supervisor option
+
+# v0.0.0
+
+* See changes on [Every PR](https://github.com/capatazlib/go-capataz/pulls?q=is%3Apr+is%3Aclosed+label%3Apre-changelog) that was created before a CHANGELOG file was added
