+++ cap/err_tolerance.go
@@ -1,4 +1,4 @@
-package c
+package cap
 
 import (
 	"time"
