+++ cap/err_tolerance_internal_test.go
@@ -1,4 +1,4 @@
-package c
+package cap
 
 import (
 	"testing"
