+++ cap/spec.go
@@ -100,7 +100,7 @@ type BuildNodesFn = func() ([]Node, CleanupResourcesFn, error)
 // siblings as well) when the node fails in unexpected ways.
 type SupervisorSpec struct {
 	name            string
-	errTolerance    c.ErrTolerance
+	errTolerance    ErrTolerance
 	buildNodes      BuildNodesFn
 	order           Order
 	strategy        Strategy
@@ -249,7 +249,7 @@ func NewSupervisorSpec(name string, buildNodes BuildNodesFn, opts ...Opt) Superv
 		// Children will have a tolerance of 1 error every 5 seconds before telling
 		// the supervisor to give up, this is insipired by Erlang OTP documentation.
 		// http://erlang.org/doc/design_principles/sup_princ.html#maximum-restart-intensity
-		errTolerance:    c.ErrTolerance{MaxErrCount: 1, ErrWindow: 5 * time.Second},
+		errTolerance:    ErrTolerance{MaxErrCount: 1, ErrWindow: 5 * time.Second},
 		buildNodes:      buildNodes,
 		shutdownTimeout: defaultSupShutdownTimeout,
 		eventNotifier:   emptyEventNotifier,
